move files with directory structure
This script uses the System.Windows.Forms.FolderBrowserDialog to interactively prompt the user to select the source and destination directories. It then proceeds to move files created on a specified target date from the source to the destination directory. If the user cancels the operation at any point, a corresponding message is displayed.