Create move_the_files_without_source_directory_structure
This script uses System.Windows.Forms.FolderBrowserDialog to prompt the user to select the destination directory and then proceeds to move files created on a specified target date from the source directory to the selected destination directory. If the user cancels the operation at any point, a corresponding message is displayed.